Exploring interactivity detection
The file `test` in this commit represents a proto
-type of what checking for an interactive execution
might look like. The goal is to confirm if this
would work on all flavors of Linux (the part that
is likely to require refining is the location of
the `stdout` descriptor/device).

The reason I am committing this mid-experimentation
is to steer back from trying to fetch and filter
tags, while looking for suitable architectures, as
I think this is turning out to be much more complex
than I initially anticipated, and borders ridiculous
given what I am trying to achieve. Hence, after this
commit, I want to go back and clean up the code so
that:
- I do not attempt to fetch any specific tag for
  any of the test images, instead, I just use what
  ever have been fetched earlier.
- If no tags were found locally then I try to fetch
  the 'latest' tag which exists on most images.
- If this fails, either due to the image not having
  a 'latest' tag or because it is not available for
  the current architecture, then we skip the test
  image and remove it from the test-image array.