Saving ongoing work before changing direction
So, as I was not able to run the unit test commands
and capture their output without changing their
behavior, I am finding myself in a place where
I am forced to use `docker run` for each one of the
commands. As far as I can think, this would be the
only right way to do this, without more knowledge
in Linux than what I currently have.

This requires a major architecture change as docker
moves from being an orchestrator of the testing of
the whole unit, to just being a tool or an executor
for each unit test command, implying that it would
have to be called multiple times for each unit.

From a performance standpoint this is a negative
point but it would allow me to hopefully execute
this correctly.

Speaking of changes, the following are needed:
- A script is still needed to install 'bash' as
  a dependency for some of the docker images.
- This setup is recommended to be carried through
  a script independent from the unit, which in this
  case called `prototype`.
- The setup script becomes the entrypoint script,
  which means it has to orchestrate or call the
  single unit-test command somehow after bash is
  installed (could be passed as an argument?)
- The main test script, or the test orchestrator,
  is now responsible of parsing the test commands
  and spinning up containers, as well as comparing
  the command outcome with the desired outcome.
- In summary, the test main script lives on the
  machine, while the setup script and the test-
  command executer spawned from it have to be copied
  to the Docker container used as executor.
- Below are some intuitive names for the new
  components which I will try to follow in the next
  commit:
  - `run-tests.sh` (outside container)
  - `test-spec.sh` (outside container)
  - `test-entrypoint.sh` (copied to container)
  - `test-command.sh`    (copied to container)