Added local testing capability and cleaned up branch
The unit testing shell script can now be called in non containerised
environment which required that it be passed an arbitrary path to
serve as the path for the executable under test.

I also realised I can use test.log momentarily and delete it after
wards, while returning its content as part of a single unit testing
step/instruction. This simplified the task of checking the expected
outcome from all the steps.