Fixed many bugs in the logic of the cli tool
The changes are based on a better analysis and understanding of the
different scenarios and calling cases. Namely:

1. case 1/3) the cli tool is called interactively, in which case:
  - docker needs to be available
  - a docker image need to be sufficiently specified
  - an output archive need to be provided
2. case 2/3) the cli tool is called to generate an eval string:
  - a docker image need to be sufficiently specified, only
3. case 3/3) the cli tool is called through an eval string:
  - docker needs to be available
  - a docker image need to be sufficiently specified and passed

Additionally, the cli tool can be called with the help '--help'
option, as well.

If in any of these cases, some of the conditions are not met,
then the call is incomplete. If the call is incomplete, the
help message for that scenario need to be displayed. if the
help option is provided, on the other hand, then both have
to be displayed.

From this understanding and analysis, I was able to have much
more control over the behavior of the tool, and now I have
slightly more confidence that it will work in most situations.