Added debug support
I am still facing issues having full code coverage, which is ideally
achieved through tests.

The execution path I am unable to get into right now is the one
triggered when the tool determines that the remote SSH invocation
is incomplete and hence must be returned from with failure becase:
- a noninteractive execution prohibits asking for user input
- a consent is needed to download a docker image on the remote

Note that this desire for content is particularly true given that
many of these images are large enough to deplete space, and that
one is probably using VMs with limited disk space for building and
exporting LXC templates.