The cli tool now works, for the first time
The code is absolutely horrible but I am approaching
this tool like an artist, not a programmer.

By this I mean, that I am starting by crafting what
I want to become, a very nice user experience, while
also figuring out along the way what I (and others)
might want to achieve using it.

This is a slightly different attitude and a some
what different approach that what you would expect
from a technical person, but this is how I am
approaching it.