Articulate goals better in README.me and fix a bug
The bug fixed had to do with a malfunction when
forwarding output in the second usage scenario.

This was due to prompts asking for user input being
printed to stdout, which was being forwarded to a
file in this case. The issue was solved by forwarding
these prompts to stderr instead.

This also helped discover similar bugs related to
the same issue, namely forgetting to relay the
output of printf and echo to stderr.

Finally, with the help of an LLM (DeepSeek R1) I
was able to better articulate why I am publishing
this code and why I think it goes a little beyond
just a helper cli for Proxmox. This led to what I
believe to be a better text than what I was able to
achieve on my own (being in my own head, as anyone
who spent 3 days straight trying to figure the nitty
gritty details of how to make a concept work in
reality, while not sure it was possible in the first
place).